requirements.txt
click==6.6
-e git+git@github.com:dhilipsiva/pwdman.git@63c2dcc164a428f2fa2985ffdd081dd47766efd5#egg=pwdman
pycrypto==2.6.1
